Properly handle both v1 and v2 span-ids.
The SPAN_ID is a decimal number in v1 of the trace-context API,
but a (16-digit) hexadecimal number in v2:
   https://cloud.google.com/python/docs/reference/cloudtrace/1.2.0/google.cloud.trace_v1.types.TraceSpan
   https://cloud.google.com/python/docs/reference/cloudtrace/1.2.0/google.cloud.trace_v2.types.Span

Stackdriver uses only v2 these days:
   https://issuetracker.google.com/issues/338634230?pli=1

Unfortunately, the X-Cloud-Trace-Context may still be using the
v1 API.  We try to detect that situation and convert it to a v2
compatible trace-id by formatting the number in hex.

We cannot perfectly detect whether we're v1 or v2, because
there are some decimal numbers that look like 16-digit hex
numbers.  But not very many; we'd expect the logic below to
make a mistake about 1 in every 2,000 log entries.  In such
cases, we will use the wrong span-id, making it harder to
associate log-messages with the requestlog that spawned it.

Issue: https://console.cloud.google.com/support/cases/detail/v2/53379415?project=khan-academy

Test plan:
Fingers crossed, it can't be any worse than what we have now!